employe controllers done

+++ src/models/employesModel.ts
@@ -1,27 +1,34 @@
 import employeModel from "../db/dbModels/employeModel";
-import { TEmployeBase, TFilterParams, TUpdateEmployeParams } from "../utils/types/employes.interface";
+import { IModelClass, TEmployeBase, TFilterParams, TUpdateEmployeParams } from "../utils/types/employes.interface";
 
-export default class EmployeModel {
-  static async getAll():Promise<TEmployeBase[] | []>{
+
+
+export default class EmployeModel implements IModelClass {
+  async getAll():Promise<TEmployeBase[] | []>{
     const employes = await employeModel.find()
     return employes
   }
 
-  static async getById(id:string): Promise<TEmployeBase>{
+  async getById(id:string): Promise<TEmployeBase>{
     const employe = await employeModel.findById(id)
     if(!employe) throw new TypeError('employe with that ID does not exists')
     else return employe
   }
 
-  static async getByName({name, lastname}: {name: string, lastname: string}):Promise<TEmployeBase>{
-  const employe = await employeModel.findOne({name, lastname})
-  if(!employe) throw new TypeError('employe with that NAME and LASTNAME does not exists')
-  else return employe
+  async getByName({name, lastname}: {name: string, lastname: string}):Promise<TEmployeBase>{
+    const employe = await employeModel.findOne({name, lastname})
+    if(!employe) throw new TypeError('employe with that NAME and LASTNAME does not exists')
+    else return employe
   }
 
-  static async filterByProp({type, value}: TFilterParams): Promise<TEmployeBase[]>{
+  async filterByProp({type, value}: TFilterParams): Promise<TEmployeBase[]>{
+    if(!type || !value) throw new TypeError('missing values of TYPE and VALUE props ')
     if(type === 'position'){
       const employes: TEmployeBase[] = await employeModel.find({position: value})
+      return employes
+    }
+    if(type === 'area'){
+      const employes: TEmployeBase[] = await employeModel.find({area: value})
       return employes
     }
     if(type === 'gender'){
@@ -32,26 +39,30 @@ export default class EmployeModel {
       const employes: TEmployeBase[] = await employeModel.find({nationality: value})
       return employes
     }
-    else return []
+    else throw new TypeError('type is needed for filtering')
   }
 
-  static async createEmploye(data: TEmployeBase): Promise<TEmployeBase>{
+  async createEmploye(data: TEmployeBase): Promise<TEmployeBase>{
     const newEmploye: TEmployeBase = await employeModel.create(data)
+    if(!newEmploye) throw new TypeError('an error ocurren, employe not created')
     return newEmploye
   }
 
-  static async updateEmploye({id, data}: {id: string, data: TUpdateEmployeParams }){
+  async updateEmploye({id, data}: {id: string, data: TUpdateEmployeParams }):Promise<TEmployeBase|null>{
+    if(!id || !data ) throw new TypeError('missing values')
     const updateEmploye = await employeModel.findByIdAndUpdate(id, {data}, {new: true})
+    if(!updateEmploye) throw new TypeError('employed not found')
     await updateEmploye?.save()
     return updateEmploye
   }
 
-  static async removeEmployeById({id}:{id: string}){
+  async removeEmployeById({id}:{id: string}):Promise<void>{
+    if(!id) throw new TypeError('ID is required')
     await employeModel.deleteOne({id})
   }
 
-  static async removeManyFromSector({position}: {position: string}){
-    await employeModel.deleteMany({position})
+  async removeManyFromArea({area}: {area: string}):Promise<void>{
+    await employeModel.deleteMany({area})
   }
 
 }
